+++ lib/Options.php
@@ -17,13 +17,16 @@ class Options {
 	/**
 	 * Register hooks.
 	 *
+	 * @since Unreleased Register the update action only if the user is in the admin.
 	 * @since 0.5.0 Added conditions for registering the update action.
 	 * @since 0.0.11
 	 *
 	 * @return null
 	 */
 	public function register() : void {
-		\add_action( 'wp_loaded', [ self::class, 'update' ] );
+		if ( \is_admin() ) {
+			\add_action( 'wp_loaded', [ self::class, 'update' ] );
+		}
 	}
 
 	/**